Add meta viewport to disable device scaling
A shorthand to disable device scaling, or to use some other option on it if customization is needed. Notably nearly all advanced sites appear to be using the setting that is obtained by `Document(_viewport=True)`. We do not still make that the default, as using the option means having to write CSS for responsive web design.